requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
cloudinary==1.44.2
Django==6.0.4
django-cloudinary-storage==0.3.0
idna==3.11
Markdown==3.10.2
python-dotenv==1.2.2
requests==2.33.1
six==1.17.0
sqlparse==0.5.5
tzdata==2026.1
urllib3==2.6.3
whitenoise==6.12.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
psycopg[binary]
dj-database-url==2.3.0
